src_api_gateway/web_service/requirements.txt
flask
Flask-Session
redis
flask-login
requests
boto3
python-dotenv
Flask-Dance
Authlib
